Most of mapMgr, and more (untested)
Entity mutex manipulation, may want to decouple camera and entity,
position hash function, most of the entity and chunk management
functions in map manager. NOTHING IS TESTED

+++ code/engine/map/entity/entity.cpp
@@ -36,7 +36,7 @@
 
 	Assigns values, starts with no components
 */
-entity::entity(uint64 _UID, uint32 time)
+entity::entity(uint32 _UID, uint32 time)
 {
 	UID = _UID;
 	lastUpdate = time;
